Map and Tile Classes
Created the Map and Tile classes.

Still need to test a few things, namely the SetTile functionality.
Just need to make sure the references copy over just fine. New to C#, so
hopefully it works the way how I intended...

Also a Godot script is connected to the map, which is titled
GodotMapScript. This shall manage the MapGenerator to call the generate
stuff.